requirements.txt
appdirs==1.4.4
asgiref==3.6.0
Backtesting==0.3.3
backtrader==1.9.78.123
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clr-loader==0.2.5
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.3
cssselect2==0.7.0
distlib==0.3.6
dj-database-url==1.3.0
Django==4.2
django-heroku==0.3.1
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.41.1
frozendict==2.3.8
html5lib==1.1
idna==3.4
inflection==0.5.1
IronPdf==2023.7.9
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.2
more-itertools==9.1.0
multitasking==0.0.11
mysql-connector-python==8.0.32
Nasdaq-Data-Link==1.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parglare==0.16.1
pdfkit==1.0.0
Pillow==10.0.0
platformdirs==3.3.0
protobuf==3.20.3
psycopg2==2.9.6
pycparser==2.21
pydyf==0.7.0
pyOpenSSL==23.2.0
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonnet==3.0.1
pytz==2023.3
PyYAML==6.0
requests==2.28.2
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tinycss2==1.2.1
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.22.0
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.3.4
whitenoise==6.4.0
xyzservices==2023.5.0
yfinance==0.2.22
zopfli==0.2.2
